Release 005

+++ src/main/java/ru/skypro/homework/service/impl/AdsServiceImpl.java
@@ -6,12 +6,15 @@ import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import ru.skypro.homework.dto.*;
+import ru.skypro.homework.dto.AdsDto;
+import ru.skypro.homework.dto.CreateAdsDto;
+import ru.skypro.homework.dto.FullAdsDto;
+import ru.skypro.homework.dto.ResponseWrapperAds;
 import ru.skypro.homework.entity.Ads;
 import ru.skypro.homework.entity.Image;
+import ru.skypro.homework.entity.User;
 import ru.skypro.homework.exceptions.AdsNotFoundException;
 import ru.skypro.homework.mapper.AdsMapper;
-import ru.skypro.homework.mapper.UserMapper;
 import ru.skypro.homework.repository.AdsRepository;
 import ru.skypro.homework.service.AdsService;
 import ru.skypro.homework.service.ImageService;
@@ -25,16 +28,14 @@ import java.util.List;
 public class AdsServiceImpl implements AdsService {
     private final AdsRepository adsRepository;
     private final AdsMapper adsMapper;
-    private final UserMapper userMapper;
     private final ImageService imageService;
     private final UserService userService;
 
     /**
-     * Receive all Ads
-     * The Service method is being used{@link AdsService#getAllAds()}
-     * wherein the repository method is used to get all declarations{@link AdsRepository#findAll()}
+     * Receive all Ads.
+     * The repository method {@link AdsRepository#findAll()} is used to get a List of all ads.
      *
-     * @return ad list
+     * @return {@link ResponseWrapperAds} instance with number of founded ads and List of {@link AdsDto}
      */
     @Override
     public ResponseWrapperAds getAllAds() {
@@ -44,20 +45,20 @@ public class AdsServiceImpl implements AdsService {
     }
 
     /**
-     * Creating of new ad
+     * Creates new ad.
      *
-     * @param createAdsDto
-     * @param image
-     * @param authentication
-     * @return ad created
+     * @param createAdsDto {@link CreateAdsDto} from a client
+     * @param image {@link MultipartFile} with image from a client
+     * @param authentication {@link Authentication} instance from controller
+     * @return {@link AdsDto} instance of the created ad
      */
     @Override
     @Transactional
     public AdsDto createAds(CreateAdsDto createAdsDto, MultipartFile image, Authentication authentication) {
         log.info("Was invoked createAds method from {}", AdsService.class.getSimpleName());
-        UserDto currentUserDto = userService.getUserByEmail(authentication.getName());
+        User currentUser = userService.getUser(authentication.getName());
         Ads ads = adsMapper.createAdsDtoToAds(createAdsDto);
-        ads.setAuthor(userMapper.userDtoToUser(currentUserDto));
+        ads.setAuthor(currentUser);
         Ads savedAds = adsRepository.save(ads);
 
         Image adsImage = imageService.createImage(image, savedAds);
@@ -66,10 +67,10 @@ public class AdsServiceImpl implements AdsService {
     }
 
     /**
-     * Search for a full ad in DB by ID
+     * Search for a full ad in DB by ID.
      *
-     * @param id
-     * @return
+     * @param id identification number of ad
+     * @return {@link FullAdsDto} instance for desired ad
      */
     @Override
     public FullAdsDto getFullAdsById(long id) {
@@ -79,35 +80,34 @@ public class AdsServiceImpl implements AdsService {
     }
 
     /**
-     * Delete ad from DB by id
-     * The repository method is being used {@link AdsRepository#delete(Object)}
+     * Delete ad from DB by id.
+     * The repository method {@link AdsRepository#delete(Object)} is used.
      *
-     * @param id
-     * @param authentication
+     * @param id identification number of ad
+     * @param authentication {@link Authentication} instance from controller
      */
     @Override
     public void removeAds(long id, Authentication authentication) {
         log.info("Was invoked removeAds method from {}", AdsService.class.getSimpleName());
         Ads ads = getAdsById(id);
 
-        checkIfUserCanAlterAds(authentication, ads);
+        userService.checkIfUserHasPermissionToAlter(authentication, ads.getAuthor().getUsername());
         adsRepository.delete(ads);
     }
 
     /**
-     * Receive old ad by id, update and save
+     * Update an ad by its ID.
      *
-     * @param id
-     * @param createAdsDto
-     * @param authentication
-     * @return ad update
+     * @param id identification number of ad
+     * @param createAdsDto {@link CreateAdsDto} instance from a client
+     * @param authentication {@link Authentication} instance from controller
+     * @return {@link AdsDto} instance with updated information
      */
     @Override
     public AdsDto updateAdsById(long id, CreateAdsDto createAdsDto, Authentication authentication) {
         log.info("Was invoked updateAdsById method from {}", AdsService.class.getSimpleName());
         Ads oldAds = getAdsById(id);
-
-        checkIfUserCanAlterAds(authentication, oldAds);
+        userService.checkIfUserHasPermissionToAlter(authentication, oldAds.getAuthor().getUsername());
         Ads infoToUpdate = adsMapper.createAdsDtoToAds(createAdsDto);
 
         oldAds.setPrice(infoToUpdate.getPrice());
@@ -118,34 +118,24 @@ public class AdsServiceImpl implements AdsService {
         return adsMapper.adsToAdsDto(updatedAds);
     }
 
-    private void checkIfUserCanAlterAds(Authentication authentication, Ads ads) {
-        boolean matchUser = authentication.getName().equals(ads.getAuthor().getEmail());
-        boolean userIsAdmin = userService.checkIfUserIsAdmin(authentication);
-
-        if (!(userIsAdmin || matchUser)){
-            log.warn("Current Ads isn't of authentication user!");
-            throw new RuntimeException("403 Forbidden");
-        }
-    }
-
     /**
-     * Receive all ads for user
+     * Receive all ads for the particular user.
      *
-     * @param username
-     * @return ad list
+     * @param username name of a user
+     * @return {@link ResponseWrapperAds} instance with number of founded ads and List of {@link AdsDto}
      */
     @Override
     public ResponseWrapperAds getAllAdsForUser(String username) {
         log.info("Was invoked getAllAdsForUser method from {}", AdsService.class.getSimpleName());
-        List<Ads> userAdsList = adsRepository.findAdsByAuthorEmail(username);
+        List<Ads> userAdsList = adsRepository.findAdsByAuthorUsername(username);
         return adsMapper.adsListToResponseWrapperAds(userAdsList.size(), userAdsList);
     }
 
     /**
-     * Receive ad by id
-     * The repository method is being used {@link AdsRepository#findById(Object)}
-     * @param id
-     * @return ad by id
+     * Receive ad by id.
+     * The repository method {@link AdsRepository#findById(Object)} is used.
+     * @param id identification number of ad
+     * @return {@link Ads} instance
      * @throws AdsNotFoundException if no ad was found
      */
     @Override
